# scripts/deployment/openapi.yaml
openapi: 3.1.0
info:
  title: The New Fuse API
  version: 1.0.0
  description: |
    Comprehensive API documentation for The New Fuse platform - a multi-agent orchestration system with blockchain integration, workflow automation, and AI-powered capabilities.

    ## Features
    - **Agent Management**: Create, manage, and orchestrate AI agents
    - **Workflow Engine**: Build and execute complex workflows
    - **Chat System**: Real-time messaging and communication
    - **Web3 Integration**: Wallet management, transactions, and smart accounts (ERC-4337)
    - **MCP Protocol**: Model Context Protocol for agent communication
    - **Authentication**: JWT-based secure authentication
    - **Rate Limiting**: Tiered rate limiting based on user roles

    ## Authentication
    Most endpoints require JWT authentication. Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Base URLs
    - Development: `http://localhost:3001/api`
    - API Gateway: `http://localhost:4000/api/v1`
    - Production: `https://api.thenewfuse.com/api`

  contact:
    name: The New Fuse API Support
    url: https://thenewfuse.com/support
    email: support@thenewfuse.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development API Server
  - url: http://localhost:4000/api/v1
    description: API Gateway (Development)
...
    description: Production API Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from /auth/login
  schemas:
...
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
    RegisterDto:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          example: newuser@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          example: John
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          example: Doe
        phone:
          type: string
          format: phone
          example: "+1234567890"
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token (15 minutes expiry)
        refreshToken:
          type: string
          description: JWT refresh token (7 days expiry)
        user:
          $ref: "#/components/schemas/User"
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 900
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [user, admin, agent]
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time

    # Agent Schemas
...
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string

    # Pagination
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: User login
      description: Authenticate user and generate JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many requests - rate limit exceeded
  /auth/register:
    post:
      tags:
        - auth
      summary: User registration
      description: Create a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid registration data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Email already registered
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Refresh authentication token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
  /auth/logout:
    post:
      tags:
        - auth
      summary: User logout
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        "401":
          description: Not authenticated
  /auth/me:
    get:
      tags:
        - auth
      summary: Get current user
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # Resource Endpoints
...